mudança de banco

+++ back/database/migrations/2025_06_28_190231_create_att_types_table.php
@@ -9,19 +9,19 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up(): void
-    {
-        Schema::create('attachment_type', function (Blueprint $table) {
-            $table->id();
-            $table->string("name");
-            $table->integer("req_type");
-            $table->timestamps();
+   public function up(): void
+   {
+       Schema::create('attachment_type', function (Blueprint $table) {
+           $table->id();
+           $table->string("name");
+           $table->unsignedBigInteger("req_type");  // ALTERADO para unsignedBigInteger
+           $table->timestamps();
 
-            $table->foreign('req_type')
-              ->references('id')
-              ->on('request_type');
-        });
-    }
+           $table->foreign('req_type')
+                 ->references('id')
+                 ->on('request_type');
+       });
+   }
 
     /**
      * Reverse the migrations.
